refactor: move period type.
