+++ ptahlmud/core/__init__.py
@@ -1,5 +1,6 @@
 """Package core concepts."""
 
 from ptahlmud.core.fluctuations import Fluctuations
+from ptahlmud.core.period import Period
 
-__all__ = ["Fluctuations"]
+__all__ = ["Fluctuations", "Period"]
