+++ ptahlmud/datastack/clients/binance_client.py
@@ -5,9 +5,9 @@ from datetime import datetime, timedelta
 import pandas as pd
 from binance.client import Client
 
+from ptahlmud.core import Period
 from ptahlmud.core.fluctuations import Fluctuations
 from ptahlmud.datastack.clients.remote_client import RemoteClient
-from ptahlmud.types import Period
 
 
 class BinanceClient(RemoteClient):
