--- ptahlmud/types/__init__.py
@@ -1,5 +1,0 @@
-"""This package defines custom types."""
-
-from ptahlmud.types.period import Period
-
-__all__ = ["Period"]
