+++ tests/datastack/test_fluctuations_service.py
@@ -2,11 +2,10 @@ from datetime import datetime, timedelta
 
 import pytest
 
-from ptahlmud.core.fluctuations import Fluctuations
+from ptahlmud.core import Fluctuations, Period
 from ptahlmud.datastack.clients.remote_client import RemoteClient
 from ptahlmud.datastack.fluctuations_service import CustomOperation, FluctuationsService, FluctuationsSpecs
 from ptahlmud.datastack.testing.fluctuations import generate_fluctuations
-from ptahlmud.types import Period
 
 
 class MockedClient(RemoteClient):
